fix gml:AngleType in db; fix multiplicity planArt
(cherry picked from commit 221209a2ba7e2ff3b73adebcbae90be53065c281)

+++ src/SAGisXPlanung/XPlan/types.py
@@ -52,7 +52,7 @@ class RefURL(types.TypeDecorator):
 class Angle(types.TypeDecorator, XPlanungMeasureType):
     """ Repräsentation des XPlanGML AngleType """
 
-    impl = types.Integer
+    impl = types.Float
     python_type = float
     cache_ok = True
